+++ subsys/bluetooth/controller/ll_sw/lll.h
@@ -103,6 +103,9 @@ enum {
 	TICKER_ID_SCAN_SYNC_ISO_BASE,
 	TICKER_ID_SCAN_SYNC_ISO_LAST = ((TICKER_ID_SCAN_SYNC_ISO_BASE) +
 					(CONFIG_BT_CTLR_SCAN_SYNC_ISO_SET) - 1),
+	TICKER_ID_SCAN_SYNC_ISO_RESUME_BASE,
+	TICKER_ID_SCAN_SYNC_ISO_RESUME_LAST = ((TICKER_ID_SCAN_SYNC_ISO_RESUME_BASE) +
+					       (CONFIG_BT_CTLR_SCAN_SYNC_ISO_SET) - 1),
 #endif /* CONFIG_BT_CTLR_SYNC_ISO */
 #endif /* CONFIG_BT_CTLR_ADV_PERIODIC */
 #endif /* CONFIG_BT_CTLR_ADV_EXT */
