+++ subsys/bluetooth/controller/ll_sw/ull.c
@@ -127,7 +127,9 @@
 				   (TICKER_ID_SCAN_SYNC_BASE) + 1)
 #if defined(CONFIG_BT_CTLR_SYNC_ISO)
 #define BT_SCAN_SYNC_ISO_TICKER_NODES ((TICKER_ID_SCAN_SYNC_ISO_LAST) - \
-				       (TICKER_ID_SCAN_SYNC_ISO_BASE) + 1)
+				       (TICKER_ID_SCAN_SYNC_ISO_BASE) + 1 + \
+				       (TICKER_ID_SCAN_SYNC_ISO_RESUME_LAST) - \
+				       (TICKER_ID_SCAN_SYNC_ISO_RESUME_BASE) + 1)
 #else /* !CONFIG_BT_CTLR_SYNC_ISO */
 #define BT_SCAN_SYNC_ISO_TICKER_NODES 0
 #endif /* !CONFIG_BT_CTLR_SYNC_ISO */
